requirements-dev.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# MLOps & Experiment Tracking
mlflow>=2.5.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging
tensorboard>=2.13.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
ipykernel>=6.25.0
jupyter>=1.0.0


sqlalchemy>=2.0.0
#psycopg2-binary
#pymysql  

fastapi 
uvicorn 
python-multipart
onnx 
onnxscript